Update unit test to the new data structure
