+++ libs/coin-evm/src/__tests__/unit/api/gasTracker/index.unit.test.ts
@@ -8,7 +8,7 @@ const fakeCurrency: Partial<CryptoCurrency> = {
     chainId: 1,
     gasTracker: {
       type: "ledger",
-      uri: "my-gas-tracker.com",
+      explorerId: "eth",
     },
   },
   units: [{ code: "ETH", name: "ETH", magnitude: 18 }],
