+++ libs/coin-evm/src/__tests__/unit/prepareTransaction.unit.test.ts
@@ -1,6 +1,6 @@
 import BigNumber from "bignumber.js";
 import { prepareForSignOperation, prepareTransaction } from "../../prepareTransaction";
-import * as rpcAPI from "../../api/rpc/rpc.common";
+import * as nodeApi from "../../api/node/rpc.common";
 import {
   account,
   expectedData,
@@ -16,9 +16,9 @@ describe("EVM Family", () => {
   describe("prepareTransaction.ts", () => {
     beforeEach(() => {
       // These mocks will be overriden in some tests
-      jest.spyOn(rpcAPI, "getGasEstimation").mockImplementation(async () => new BigNumber(21000));
+      jest.spyOn(nodeApi, "getGasEstimation").mockImplementation(async () => new BigNumber(21000));
       // These mocks will be overriden in some tests
-      jest.spyOn(rpcAPI, "getFeesEstimation").mockImplementation(async () => ({
+      jest.spyOn(nodeApi, "getFeesEstimation").mockImplementation(async () => ({
         gasPrice: new BigNumber(1),
         maxFeePerGas: new BigNumber(1),
         maxPriorityFeePerGas: new BigNumber(1),
@@ -57,7 +57,7 @@ describe("EVM Family", () => {
 
       describe("Coins", () => {
         it("should have a gasLimit = 0 when recipient has an error", async () => {
-          jest.spyOn(rpcAPI, "getGasEstimation").mockImplementation(async () => {
+          jest.spyOn(nodeApi, "getGasEstimation").mockImplementation(async () => {
             throw new Error();
           });
 
@@ -78,7 +78,7 @@ describe("EVM Family", () => {
         });
 
         it("should have a gasLimit = 0 when amount has an error", async () => {
-          jest.spyOn(rpcAPI, "getGasEstimation").mockImplementation(async () => {
+          jest.spyOn(nodeApi, "getGasEstimation").mockImplementation(async () => {
             throw new Error();
           });
 
@@ -134,7 +134,7 @@ describe("EVM Family", () => {
         });
 
         it("should return a legacy coin transaction", async () => {
-          jest.spyOn(rpcAPI, "getFeesEstimation").mockImplementationOnce(async () => ({
+          jest.spyOn(nodeApi, "getFeesEstimation").mockImplementationOnce(async () => ({
             gasPrice: new BigNumber(1),
             maxFeePerGas: null,
             maxPriorityFeePerGas: null,
@@ -179,7 +179,7 @@ describe("EVM Family", () => {
 
         it("should do a gas estimation when data has been added to the coin transaction", async () => {
           jest
-            .spyOn(rpcAPI, "getGasEstimation")
+            .spyOn(nodeApi, "getGasEstimation")
             .mockImplementationOnce(async () => new BigNumber(12));
 
           const accountWithBalance = {
@@ -205,7 +205,7 @@ describe("EVM Family", () => {
 
       describe("Tokens", () => {
         it("should have a gasLimit = 0 and no data when recipient has an error", async () => {
-          jest.spyOn(rpcAPI, "getGasEstimation").mockImplementation(async () => {
+          jest.spyOn(nodeApi, "getGasEstimation").mockImplementation(async () => {
             throw new Error();
           });
 
@@ -227,7 +227,7 @@ describe("EVM Family", () => {
         });
 
         it("should have a gasLimit = 0 when amount has an error", async () => {
-          jest.spyOn(rpcAPI, "getGasEstimation").mockImplementation(async () => {
+          jest.spyOn(nodeApi, "getGasEstimation").mockImplementation(async () => {
             throw new Error();
           });
 
@@ -250,7 +250,7 @@ describe("EVM Family", () => {
 
         it("should create a token transaction using all amount in the token account", async () => {
           jest
-            .spyOn(rpcAPI, "getGasEstimation")
+            .spyOn(nodeApi, "getGasEstimation")
             .mockImplementationOnce(async () => new BigNumber(12));
 
           const tokenAccountWithBalance = {
@@ -285,7 +285,7 @@ describe("EVM Family", () => {
 
         it("should go to gas estimation with a transaction without 0 amount", async () => {
           jest
-            .spyOn(rpcAPI, "getGasEstimation")
+            .spyOn(nodeApi, "getGasEstimation")
             .mockImplementationOnce(async () => new BigNumber(12));
 
           const tokenAccountWithBalance = {
@@ -303,7 +303,7 @@ describe("EVM Family", () => {
             subAccountId: tokenAccountWithBalance.id,
           });
 
-          expect(rpcAPI.getGasEstimation).toBeCalledWith(
+          expect(nodeApi.getGasEstimation).toBeCalledWith(
             account2,
             expect.objectContaining({
               recipient: tokenAccount.token.contractAddress,
@@ -337,7 +337,7 @@ describe("EVM Family", () => {
         });
 
         it("should return a legacy token transaction", async () => {
-          jest.spyOn(rpcAPI, "getFeesEstimation").mockImplementationOnce(async () => ({
+          jest.spyOn(nodeApi, "getFeesEstimation").mockImplementationOnce(async () => ({
             gasPrice: new BigNumber(1),
             maxFeePerGas: null,
             maxPriorityFeePerGas: null,
@@ -375,7 +375,7 @@ describe("EVM Family", () => {
               feesStrategy: undefined,
             });
 
-            expect(rpcAPI.getFeesEstimation).toBeCalledTimes(1);
+            expect(nodeApi.getFeesEstimation).toBeCalledTimes(1);
 
             expect(tx).toEqual({
               ...transaction,
@@ -396,7 +396,7 @@ describe("EVM Family", () => {
               feesStrategy: "custom",
             });
 
-            expect(rpcAPI.getFeesEstimation).toBeCalledTimes(0);
+            expect(nodeApi.getFeesEstimation).toBeCalledTimes(0);
 
             expect(tx).toEqual({
               ...transaction,
@@ -415,7 +415,7 @@ describe("EVM Family", () => {
               gasOptions: undefined,
             });
 
-            expect(rpcAPI.getFeesEstimation).toBeCalledTimes(1);
+            expect(nodeApi.getFeesEstimation).toBeCalledTimes(1);
 
             expect(tx).toEqual({
               ...transaction,
@@ -452,7 +452,7 @@ describe("EVM Family", () => {
               gasOptions,
             });
 
-            expect(rpcAPI.getFeesEstimation).toBeCalledTimes(0);
+            expect(nodeApi.getFeesEstimation).toBeCalledTimes(0);
 
             expect(tx).toEqual({
               ...transaction,
@@ -469,7 +469,7 @@ describe("EVM Family", () => {
 
       describe("Nfts", () => {
         it("should have a gasLimit = 0 and no data when recipient has an error", async () => {
-          jest.spyOn(rpcAPI, "getGasEstimation").mockImplementation(async () => {
+          jest.spyOn(nodeApi, "getGasEstimation").mockImplementation(async () => {
             throw new Error();
           });
 
@@ -495,7 +495,7 @@ describe("EVM Family", () => {
         });
 
         it("should have a gasLimit = 0 when amount has an error", async () => {
-          jest.spyOn(rpcAPI, "getGasEstimation").mockImplementation(async () => {
+          jest.spyOn(nodeApi, "getGasEstimation").mockImplementation(async () => {
             throw new Error();
           });
 
@@ -541,7 +541,7 @@ describe("EVM Family", () => {
           });
 
           it("should return a legacy erc721 nft transaction", async () => {
-            jest.spyOn(rpcAPI, "getFeesEstimation").mockImplementationOnce(async () => ({
+            jest.spyOn(nodeApi, "getFeesEstimation").mockImplementationOnce(async () => ({
               gasPrice: new BigNumber(1),
               maxFeePerGas: null,
               maxPriorityFeePerGas: null,
@@ -589,7 +589,7 @@ describe("EVM Family", () => {
           });
 
           it("should return a legacy erc1155 nft transaction", async () => {
-            jest.spyOn(rpcAPI, "getFeesEstimation").mockImplementationOnce(async () => ({
+            jest.spyOn(nodeApi, "getFeesEstimation").mockImplementationOnce(async () => ({
               gasPrice: new BigNumber(1),
               maxFeePerGas: null,
               maxPriorityFeePerGas: null,
@@ -620,7 +620,7 @@ describe("EVM Family", () => {
 
     describe("prepareForSignOperation", () => {
       beforeEach(() => {
-        jest.spyOn(rpcAPI, "getTransactionCount").mockImplementation(() => Promise.resolve(10));
+        jest.spyOn(nodeApi, "getTransactionCount").mockImplementation(() => Promise.resolve(10));
       });
       afterEach(() => {
         jest.restoreAllMocks();
