+++ libs/coin-evm/src/__tests__/unit/signOperation.unit.test.ts
@@ -7,14 +7,17 @@ import { buildSignOperation, applyEIP155 } from "../../signOperation";
 import { EvmAddress, EvmSignature, EvmSigner } from "../../signer";
 import { Transaction as EvmTransaction } from "../../types";
 import { makeAccount } from "../fixtures/common.fixtures";
-import * as rpcAPI from "../../api/rpc/rpc.common";
+import * as nodeApi from "../../api/node/rpc.common";
 import { getEstimatedFees } from "../../logic";
 
 const currency: CryptoCurrency = {
   ...getCryptoCurrencyById("ethereum"),
   ethereumLikeInfo: {
     chainId: 1,
-    rpc: "my-rpc.com",
+    node: {
+      type: "external" as const,
+      uri: "my-rpc.com",
+    },
   },
 };
 const account: Account = makeAccount(
@@ -76,7 +79,7 @@ describe("EVM Family", () => {
   describe("signOperation.ts", () => {
     describe("signOperation", () => {
       beforeAll(() => {
-        jest.spyOn(rpcAPI, "getTransactionCount").mockImplementation(async () => 1);
+        jest.spyOn(nodeApi, "getTransactionCount").mockImplementation(async () => 1);
       });
 
       afterAll(() => {
@@ -111,6 +114,8 @@ describe("EVM Family", () => {
               accountId: account.id,
               transactionSequenceNumber: 1,
               date: expect.any(Date),
+              subOperations: [],
+              nftOperations: [],
               extra: {},
             });
             expect(signature).toBe(
